backend/requirements.txt
Django==6.0.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
Pillow==10.2.0
cloudinary==1.40.0
django-cloudinary-storage==0.3.0
razorpay==1.4.2
psycopg2-binary==2.9.9